requirements.txt
fastapi[standard]
python-dotenv
sqlmodel
pydantic_settings
passlib
PyJWT
passlib[bcrypt]
psycopg2-binary
qrcode[pil]
reportlab
minio
httpx
pytest
faker

